Module pegassas_aggregator::pegassas_aggregator::parser

Structs

Battery
Container structure for battery informations

BatteryData
Battery probe information structure

Cell
Container structure for cell informations

CellData
Cell probe information structure

Charge
A charge reading with a unit

Current
A current reading with a unit

Data
Top level structure to store deserialized data from probes

Frequency
A frequency reading with a unit

Power
A power reading with a unit

Solar
Container structure for Solar panel informations

SolarData
Solar Panel information structure

Temperature
A temperature reading with a unit

Voltage
A voltage reading with a unit

Enums

ChargeUnit
Charge metric

CurrentUnit
Current metric

FrequencyUnit
Frequency metric

PowerUnit
Power unit metric

TemperatureUnit
Temperature metric

VoltageUnit
Voltage metric

Functions

parse_data
Parse information from data probe and generate point to be added to Influx database